move config from ledger to lib folder
